week03-homework/pyproject.toml
[project]
name = "week03-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"python-dotenv",
	# add your dependencies here and run `uv sync` to install them
	"llama-index",
	"llama-index-core",
	"llama-index-llms-openai-like",
	"llama-index-llms-dashscope",
	"llama-index-embeddings-dashscope",
	"llama-index-llms-ollama",
	"tabulate",
	"paddlepaddle",
	"paddleocr"
]
